+++ wowusky/__init__.py
@@ -1,4 +1,4 @@
 """wowusky — minimalist World of Warcraft addon manager for Linux."""
 
-__version__ = "0.9.3"
+__version__ = "0.9.4"
 APP_NAME    = "wowusky"